Fixed php-fpm package name resolution for pacman.
For arch based distros, there have always been one official version of
php-fpm which is always pointing the latest php version. as of the time
this commit was written, php-fpm is currently at version 8.4.

The package name will remain thesame while the version will change
accordingly. So this commit is just to help pacman users have the
appropriate name for the php-fpm version.